requirements.txt
python-telegram-bot>=21.0
anthropic>=0.40.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
httpx>=0.27.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
youtube-transcript-api>=0.6.0
yt-dlp>=2024.0.0
